+++ app/controllers/sail/application_controller.rb
@@ -9,5 +9,13 @@ module Sail
     def current_user
       main_app.scope.request.env["warden"]&.user
     end
+
+    def default_url_options
+      { locale: I18n.locale }
+    end
+
+    def set_locale
+      I18n.locale = params[:locale] || I18n.default_locale
+    end
   end
 end